+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Next
+## [0.9.3](https://github.com/streamingfast/substreams-ethereum/releases/tag/v0.9.3)
 
 * Added `timestamp` on `Block` type, returns a reference to the header's timestamp.